fastapi/requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
asttokens==3.0.0
attrs==24.3.0
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.12.14
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
comm==0.2.2
debugpy==1.8.12
decorator==5.1.1
distro==1.9.0
executing==2.2.0
fastapi==0.115.8
frozenlist==1.5.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
ipykernel==6.29.5
ipython==8.31.0
jedi==0.19.2
jiter==0.8.2
jupyter_client==8.6.3
jupyter_core==5.7.2
matplotlib-inline==0.1.7
multidict==6.1.0
nest-asyncio==1.6.0
numpy==2.2.2
openai==0.28.0
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
parso==0.8.4
platformdirs==4.3.6
prompt_toolkit==3.0.50
propcache==0.2.1
psutil==6.1.1
pure_eval==0.2.3
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
PyMySQL==1.1.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
pyzmq==26.2.1
redis==5.2.1
requests==2.32.3
selenium==4.28.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.38
stack-data==0.6.3
starlette==0.45.3
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
trio==0.28.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.34.0
wcwidth==0.2.13
webdriver-manager==4.0.2
websocket-client==1.8.0
wsproto==1.2.0
yarl==1.18.3
